argb_ctrl/argb-test/pyproject.toml
[project]
name = "argb-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyserial>=3.5",
]
